+++ js/main.js
@@ -1,5 +1,5 @@
-import { createPhotos } from './create-photos.js';
+// import {somePhotos} from './create-photos.js';
+import './create-thumbnails.js';
+import './create-photos.js';
 
-const somePhotos = createPhotos(25);
-somePhotos();
-
+// const somePhotos = createPhotos(25);
